add 'ready' and 'done' fields to 'Order' model

+++ src/dj/app_menu/models.py
@@ -5,6 +5,6 @@ class MenuItem(models.Model):
     """Розширює MenuItemGeneral більш конкретними даними"""
     title = models.CharField(max_length=255, verbose_name='Назва')
     description = models.TextField(verbose_name='Опис')
-    price = models.IntegerField(verbose_name='Ціна')
+    price = models.FloatField(verbose_name='Ціна')
